+++ src/main/java/subway/config/FareConfiguration.java
@@ -3,11 +3,11 @@ package subway.config;
 import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import subway.domain.fare.AgeDiscountPolicy;
-import subway.domain.fare.DistanceFarePolicy;
 import subway.domain.fare.FareCalculator;
-import subway.domain.fare.FarePolicies;
-import subway.domain.fare.LineFarePolicy;
+import subway.domain.fare.policy.AgeDiscountPolicy;
+import subway.domain.fare.policy.DistanceFarePolicy;
+import subway.domain.fare.policy.FarePolicies;
+import subway.domain.fare.policy.LineFarePolicy;
 
 @Configuration
 public class FareConfiguration {
