+++ src/main/java/subway/domain/fare/policy/AgeDiscountPolicy.java
@@ -1,4 +1,7 @@
-package subway.domain.fare;
+package subway.domain.fare.policy;
+
+import subway.domain.fare.AgeGroup;
+import subway.domain.fare.FareInformation;
 
 public class AgeDiscountPolicy implements DiscountPolicy {
