+++ src/main/java/subway/domain/fare/policy/FarePolicy.java
@@ -1,4 +1,6 @@
-package subway.domain.fare;
+package subway.domain.fare.policy;
+
+import subway.domain.fare.FareInformation;
 
 public interface FarePolicy {
