+++ src/test/java/subway/domain/fare/policy/AgeDiscountPolicyTest.java
@@ -1,4 +1,4 @@
-package subway.domain.fare;
+package subway.domain.fare.policy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
+import subway.domain.fare.AgeGroup;
+import subway.domain.fare.FareInformation;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(ReplaceUnderscores.class)
