+++ src/test/java/subway/domain/fare/policy/FarePoliciesTest.java
@@ -1,4 +1,4 @@
-package subway.domain.fare;
+package subway.domain.fare.policy;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import subway.domain.Line;
 import subway.domain.Lines;
+import subway.domain.fare.FareInformation;
 
 
 @SuppressWarnings("NonAsciiCharacters")
